move some stuff in lib around
