+++ src/lib/analytics/index.tsx
@@ -1,5 +1,4 @@
 import TagManager from 'react-gtm-module'
-
 export const GTM_ID = process.env.GOOGLE_TAG_MANAGER_ID
 export const TAG_MANAGER_ARGS = {
   gtmId: process.env.GOOGLE_TAG_MANAGER_ID,
