+++ src/lib/drupal/getGlobalElements.test.ts
