+++ src/templates/globals/banners/facilityBanner/index.tsx
@@ -1,5 +1,5 @@
 import { useRef, useEffect, useState } from 'react'
-import { recordEvent } from '@/lib/utils/recordEvent'
+import { recordEvent } from '@/lib/analytics/recordEvent'
 import { regionBaseURL } from '@/lib/utils/helpers'
 import { VaBanner } from '@department-of-veterans-affairs/component-library/dist/react-bindings'
 import { NodeMetaInfo } from '@/types/dataTypes/drupal/node'
